typewriter overwrite fix and toggle hourly tips
I have been working with audio and put the volume holder from the main menu inside the main menu scene options tab. It is being disabled and not transfer to the other scene, however I don't think that matter because he is not being used in the main game to change volume, most things are being done by the Audio Manager.