Moving of Sprites and Scripts done
All of the final moves are now complete. There might be some in the future, however, the main organization issue is solved. I've tested a playthrough of the game and everything seems to be running okay. no missing files or sprites in the game. The main concern now is with the watcher going out of bounds, I do not know why. But I do not thin it was because of the folder movement.