Somewhat Okay
I have made a little progress on crating the new stick analog for the phone, however I can not test in the computer even though it works fine in my phone.
There is a bug where the other images from the canvas where the analog is does not currently show anything on the android build.
There is also a need to separate the hunter's tutorial puzzle to the main hunter's puzzle from the house. The single script is not handling things well. It is, however, under control and working okay now.