Tutorial beartrap modification and coin puzzle buttons
Added a few modifications to the tutorial puzzle to make the experience more clean for the player. There are a few new rules to set each state in case the player has the key or if the door puzzle was solved. On top of that, if the player completes the puzzle the next scene is called.
Added a few buttons to the coin puzzle that are suppose to be active only if the platform the game is currently being player at is an android. The buttons still need some testing.