Level Adjustments
This is the last stop before Cinemachine to the project.
The coin puzzle left side button had it's font fixed.
the second floor had most of it's items now with the correct tag as Furniture, so the blind creep does not pass through them.
The first and the second floor also got a check just in case there is a sprite that is not in it's correct z position or not tagged as Furniture.