Transition Adjusments
The bool variable that allows the player to go from one room to the other without pressing the interact button has now been updated for the entire Mansion doors but one. The second floor ladder is using a different script that needs to be addressed on a later date. That ladder leads to 3rd floor right side.