Skull puzzle now working with candles
The candles have been update to the new system and the skull puzzle has is now working fine the new system as well.
The Puzzle has been tested to ensure the right order is correct, there is only one bug left to correct. The bug involves the first light tuning on and off again depending on a certain position of the skulls.