Working on Notes
final testing on the new skull puzzle resolving a minor bug where the candle would re-light itself after completion of a certain stage.
The new objective is to complete the notes with working code where the references needed are found with ease and the text is loaded according to what is already presented in the game.