Bench is now a Prefab
The bench now works as a prefab and can be dragged into the scene. The problem now is that items are not loading in the containers. As of now I do not know what is causing the problem, this is just a fresh safe to start again in case something goes even more wrong.