Working on Inventory update and systems
The items now wait for the inventory slot to go up to close. There is bug when you enter, exit, and enter the trigger too fast . The slot drop down with items, go up, then down again but with no more items in the slot.
The system currently interact with canvas to know when to stop the player from waking, this needs to be updated to a game action for all of the canvas in the player movement script attached to the player.