Minor Bug fixes
A few bug fixes were implement to prevent the player from clicking on their item while there is no other container active.
A player can now open the inventory and it will distinguish when to open the the short item list or the full inventory. That was accomplished by making a new game action that works separate with the same journal toggle, but now it sends a value to the toggle