Some bugs solved
The item in question had a null container and it was trying to remove itself from it, for that reason there was a bug with the way items were being handles. that bug is solved now.
The only thing to solve left is to make the broken ladder 2 items and have them be in different locations.