Level Design Progress
The doors are working properly, the new container have been placed on the second floor. The new containers need to be placed on the first floor with the photo pictures, and another one in the third floor with the crucifix. After that, it's a green light for testing the of the rooms.